feat: separately, manually handled mesh
Moved JSX initialization of full mesh to
a separate component, `SiliconMesh`, so that
hopefully can do boolean operations on the separate
meshes to fix their z-fighting.

Also ended up, for `SiliconMesh`, subverting the normal `Entity`
pattern and directly implementing the adding of the meshes to the
scene using `three`. As of the time of development `solid-three`'s
Entity wasn't updating, even when `createEffect`'s callback
was being called and updating `group`. This is probably my
own misunderstanding, but got to work for now by essentially doing
the scene management directly.