feat: added orbit controls
Added `OrbitControls` to 3D viewer, allowing users to
rotate, pan, and zoom the silicon.

This reveals that the supposed blocking of diffusion
by the polysilicon is not actually being rendered, which
will likely have to also come from implementing boolean
geometry.

Additionally, there is a bug (or an incovenience?)
that zooming out too much causes the mesh to go out
of render distance and so disappear.